Just move the auxiliary module into an exploit
